clean up sign in gate methods into more modular approach
