+++ common/app/conf/switches/ABTestSwitches.scala
@@ -5,7 +5,7 @@ import conf.switches.SwitchGroup.{ABTests, Commercial}
 import org.joda.time.LocalDate
 
 trait ABTestSwitches {
-  
+
   Switch(
     ABTests,
     "ab-contributions-epic-precontribution-reminder-round-one",
@@ -118,11 +118,11 @@ trait ABTestSwitches {
 
   Switch(
     ABTests,
-    "ab-sign-in-gate-secundus",
+    "ab-sign-in-gate-tertius",
     "Test new sign in component on 2nd article view",
     owners = Seq(Owner.withGithub("coldlink"),Owner.withGithub("dominickendrick")),
-    safeState = Off,
-    sellByDate = new LocalDate(2020, 3, 31),
+    safeState = On,
+    sellByDate = new LocalDate(2020, 5, 1),
     exposeClientSide = true
   )
 }
