+++ static/src/javascripts/projects/common/modules/experiments/tests/sign-in-gate.js
@@ -1,8 +1,8 @@
 // @flow
-export const signInGateSecundus: ABTest = {
-    id: 'SignInGateSecundus',
-    start: '2019-12-20',
-    expiry: '2020-03-31',
+export const signInGate: ABTest = {
+    id: 'SignInGateTertius',
+    start: '2020-02-17',
+    expiry: '2020-05-01',
     author: 'Mahesh Makani, Dominic Kendrick',
     description:
         'Test adding a sign in component on the 2nd pageview of simple article templates, with higher priority over banners and epic, and a much larget audience size.',
