+++ static/src/javascripts/projects/common/modules/identity/sign-in-gate/component-event-helper.js
@@ -1,0 +1,31 @@
+// @flow
+import ophan from 'ophan/ng';
+
+type ABTestVariant = {
+    name: string,
+    variant: string,
+};
+
+type ComponentEventWithoutAction = {
+    component: OphanComponent,
+    value?: string,
+    id?: string,
+    abTest?: ABTestVariant,
+};
+
+// ophan helper methods
+export const submitComponentEvent = (componentEvent: OphanComponentEvent) => {
+    ophan.record({ componentEvent });
+};
+
+export const submitViewEvent = (componentEvent: ComponentEventWithoutAction) =>
+    submitComponentEvent({
+        ...componentEvent,
+        action: 'VIEW',
+    });
+
+export const submitClickEvent = (componentEvent: ComponentEventWithoutAction) =>
+    submitComponentEvent({
+        ...componentEvent,
+        action: 'CLICK',
+    });
